# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "DONE. API",
    "version" : "1.0.0",
    "description" : "Native Apple API. Email/password JWT and native Sign in with Apple. All timestamps UTC; history is server-owned. Durable activity events; APNs transport is not part of this phase."
  },
  "servers" : [ {
    "description" : "Local development; configure PUBLIC_ORIGIN for deployed links.",
    "url" : "http://127.0.0.1:8080"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "writerCapability" : {
        "description" : "43-character base64url private writer key; never a JWT and never included in the public link.",
        "type" : "http",
        "bearerFormat" : "Private write capability",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
...
      "ExtendSession" : {
        "type" : "object",
        "properties" : {
          "minutes" : {
            "format" : "int32",
            "type" : "integer",
            "minimum" : 1,
            "maximum" : 1440
          }
        },
        "additionalProperties" : false,
        "required" : [ "minutes" ]
      },
...
      "MeasuredProgram" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        },
        "additionalProperties" : false,
        "required" : [ "name" ]
      },
...
      "PageResponse_SessionView" : {
        "required" : [ "items", "page", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/SessionView"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "Problem" : {
        "required" : [ "status", "code", "detail" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
      "ProgramInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer",
            "minimum" : 0,
            "maximum" : 1440
          }
        },
        "additionalProperties" : false,
        "required" : [ "minutes" ]
      },
      "ProgramSnapshot" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isCustom" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "description" : "Immutable value snapshot. name is literal user content or a built-in label; localizationKey is optional. Measured values are separate from the original snapshot.",
        "required" : [ "name", "minutes", "isCustom" ]
      },
...
      "SessionView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "applianceId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "mode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "status" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "startedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "expectedEnd" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "completedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "collectedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "canceledAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "elapsedSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          },
          "serverTime" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "sourceProgramId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programSnapshot" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "measuredProgram" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "timingMode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isOpen" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "applianceId", "programId", "programName", "minutes", "mode", "status", "startedAt", "expectedEnd", "completedAt", "collectedAt", "canceledAt", "elapsedSeconds", "version", "serverTime", "sourceProgramId", "programSnapshot", "measuredProgram", "timingMode", "isOpen" ]
      },
...
      "StartSession" : {
        "type" : "object",
        "properties" : {
          "programId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "program" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramInput"
            }, {
              "type" : "null"
            } ]
          },
          "mode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "saveProgram" : {
            "anyOf" : [ {
              "type" : "boolean"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "anyOf" : [ {
              "format" : "int32",
              "type" : "integer",
              "minimum" : 1,
              "maximum" : 1440
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "description" : "mode also accepts the JSON alias timingMode. Preset programId may be paired with a minutes override. All appliance kinds support stopwatch; stopwatch forbids preset ID and minutes override."
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/appliances/{id}/sessions" : {
      "get" : {
        "operationId" : "SessionController_history_2992662508",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "page",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "default" : 0
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "size",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "default" : 20
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponse_SessionView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "SessionController_start_2992662508",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartSession"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "201" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{id}" : {
      "get" : {
        "operationId" : "SessionController_get_1279659017",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/cancel" : {
      "post" : {
        "operationId" : "SessionController_cancel_314447680",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/collect" : {
      "post" : {
        "operationId" : "SessionController_collect_1557172516",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/complete" : {
      "post" : {
        "operationId" : "SessionController_complete_1060241535",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/extend" : {
      "post" : {
        "operationId" : "SessionController_extend_393127904",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExtendSession"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/measured-program" : {
      "post" : {
        "operationId" : "SessionController_measured_500632515",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeasuredProgram"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{id}/repeat" : {
      "post" : {
        "operationId" : "SessionController_repeat_747640417",
        "tags" : [ "Session" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "201" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
...
  }
}